+++ src/django_elasticsearch_dsl_drf/views.py
@@ -1,4 +1,4 @@
-from .viewsets import *  # NOQA
+from .viewsets import BaseDocumentViewSet  # NOQA
 
 # This module is deprecated. Import from `viewsets` module instead.